requirements.txt
albumentations==1.4.11
matplotlib==3.7.2
matplotlib==3.7.5
numpy==1.23.5
numpy==1.24.4
onnx==1.16.1
onnx_graphsurgeon==0.5.2
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.9.0.80
paddleocr==2.7.3
paddleocr.egg==info
paddlepaddle==2.6.1
pafy==0.5.5
pandas==2.0.3
Pillow==10.4.0
PyQt5==5.15.10
PyQt5_sip==12.13.0
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.32.3
scipy==1.14.0
seaborn==0.13.2
setuptools==69.5.1
Shapely==2.0.4
thop==0.1.1.post2209072238
torch==1.12.1
torchvision==0.13.1
tqdm==4.66.4
wandb==0.17.4
